+++ server/src/engine/storage/mod.rs
@@ -61,6 +61,10 @@ pub fn repair() -> RuntimeResult<()> {
     v2::repair()
 }
 
+pub fn compact() -> RuntimeResult<()> {
+    v2::compact()
+}
+
 pub fn load(cfg: &Configuration) -> RuntimeResult<SELoaded> {
     // first determine if this is a new install, an existing install or if it uses the old driver
     if Path::new(v1::SYSDB_PATH).is_file() {