+++ server/src/engine/storage/v2/impls/mdl_journal.rs
@@ -64,6 +64,8 @@ use {
     },
 };
 
+pub type ModelAdapter = BatchAdapter<ModelDataAdapter>;
+
 #[cfg(test)]
 local! {
     static BATCH_INFO: Vec<BatchInfo> = Vec::new();